requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-environ==0.12.0
django-cors-headers==4.9.0
django-cryptography-django5==2.2
psycopg[binary]==3.2.12
celery==5.5.3
redis==7.0.1
Pillow==11.1.0
numpy==2.3.0
pandas==2.2.3
scikit-learn==1.7.2
gunicorn==23.0.0
yookassa
drf-spectacular==0.27.2
django-ratelimit==4.1.0
djangorestframework-simplejwt==5.3.1
httpx==0.27.2
asgiref==3.8.1

# AI‑микросервисы (ParkMate AI)
fastapi
uvicorn[standard]==0.32.1
